==> 3. families_as_points.Rmd <==
---
title: "families_points"
author: "Laura Hatt"
date: '2023-03-02'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-libraries, warning = FALSE}
library(tidyverse)
library(sf)
library(tmap)
library(tmaptools)
library(tigris)

path = "/Users/laurahatt/Documents/GitHub/child-care-deserts/"
```

```{r read-in-families_by_block_sf}
families_by_block_sf <- read_sf(paste0(path, "Shapefiles/families_by_block.shp"))
```

```{r create-lakeview-polygon, message=FALSE}
lakeview <- st_read(paste0(path, "Inputs/comm_areas.geojson")) %>% 
  filter(community == "LAKE VIEW") %>%
  st_transform("EPSG:3435") 

lakeview_buff <- st_buffer(lakeview, 5280*1.5)  #one and a half mile buffer
```

```{r select-blocks-in-lakeview}
lakeview_blocks <- lengths(st_intersects(families_by_block_sf, lakeview_buff)) > 0 
families_by_block_lakeview <- families_by_block_sf[lakeview_blocks,]
```

```{r create-df-where-each-row-represents-a-family}
families = list()
for (row in seq(1,nrow(families_by_block_lakeview), by=1)) {
  block <- families_by_block_lakeview[row,]
  points <- st_sample(block, size=block$tot_fam, type="random", exact=TRUE)
  points_df <- st_as_sf(data.frame(points), crs="EPSG:3435")
  families[[row]] <- points_df
  
  #progress bar
  if(row %% 100 == 0){
		cat("Finished", row, "of", 4202, "\n")
  }
}

families_as_points_lakeview = do.call(rbind, families)
```

```{r save-shapefile}
write_sf(obj=families_as_points_lakeview, dsn=paste0(path, "Shapefiles/families_as_points_lakeview.shp"))
```
