==> 4. provider_isochrones.Rmd <==
---
title: "Provider catchment areas (isochrones)"
author: "Laura Hatt"
date: '2023-03-03'
output: html_document
---

In this RMD file, I will create a polygon "catchment area" (isochrone) for each daycare provider, representing the area reachable by a 15-minute walk. Then, I will count the number of families with young children that live within each catchment area, in order to calculate the Slot-to-Population Ratio (SPR) for each daycare provider. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-packages, echo=FALSE}
library(sf)
library(tmap)
library(tidyverse)

#load developer version of osrm - there's a bug in the CRAN version
dev_lib_path <-"/Users/laurahatt/Documents/R/dev-versions/"
remotes::install_github('riatelab/osrm', force = FALSE, lib=dev_lib_path) 
library(osrm, lib.loc = dev_lib_path)

path = "/Users/laurahatt/Documents/GitHub/child-care-deserts/"
```

At this point, my analysis focuses on Lakeview. This is mainly due to computer processing constraints; otherwise I would study all of Chicago.

```{r load-data, message=FALSE}
lakeview <- st_read(paste0(path, "Inputs/comm_areas.geojson")) %>% 
  filter(community == "LAKE VIEW") %>%
  st_transform("EPSG:3435")
  
daycare_sf <- read_sf(paste0(path, "Shapefiles/daycare_locations.shp"))

fams_as_pts <- read_sf(paste0(path, "Shapefiles/families_as_points_lakeview.shp"))
```

```{r select-daycares-within-one-mile-of-lakeview, warning = FALSE}
lakeview_buff <- st_buffer(lakeview, 5280)
daycare_lakeview <- st_intersection(daycare_sf, lakeview_buff)
daycare_lakeview_4326 <- st_transform(daycare_lakeview, "EPSG:4326")
```

```{r create-15-min-walk-isochrones-around-daycare-providers}
prov_isochrones <- list()
for (row in seq(1,nrow(daycare_lakeview_4326), by=1)) {
  iso <- osrmIsochrone(loc = daycare_lakeview_4326[row,], breaks = 15, osrm.profile="foot")
  iso['PrvdrID'] <- st_set_geometry(daycare_lakeview_4326[row,], NULL)['PrvdrID']
  iso['DyCpcty'] <- st_set_geometry(daycare_lakeview_4326[row,], NULL)['DyCpcty']
  prov_isochrones[[row]] <- iso
  print(row)
}

prov_isochrones_df = do.call(rbind, prov_isochrones)
prov_isochrones_df <- st_transform(prov_isochrones_df, "EPSG:3435")
```

```{r plot-families-and-provider-isochrones}
tmap_mode("plot")

bbox_new <- st_bbox(st_buffer(lakeview, 5280*1.5)) 
xrange <- bbox_new$xmax - bbox_new$xmin 
yrange <- bbox_new$ymax - bbox_new$ymin 
#bbox_new[1] <- bbox_new[1] - (0.5 * xrange) # xmin - left
bbox_new[3] <- bbox_new[3] + (0.1 * xrange) #xmax - right
bbox_new[2] <- bbox_new[2] - (0.1 * yrange) #ymin - bottom
bbox_new <- bbox_new %>% st_as_sfc() 


map_provider_isochrones <- tm_shape(st_buffer(lakeview, 5280*1.5), bbox=bbox_new) + 
  tm_borders(col="white") +
  tm_shape(st_buffer(lakeview, 5280)) + tm_borders() +
  tm_shape(fams_as_pts) + tm_dots(col="darkred", size=0.01, alpha=0.25) +
  tm_shape(lakeview) + tm_borders(col="black", lwd=5) +
  tm_shape(prov_isochrones_df) + tm_borders(col="black", lwd=2.5) +
  tm_shape(daycare_lakeview) + tm_dots(col="darkgreen", size=0.5) +
  tm_layout(main.title = "Daycares Within One Mile of Lakeview: \nLocations, Isochrones, and Nearby Families",
            main.title.size = 1.25,
            main.title.position = "center",
            legend.position = c("right", "center")) + 
  tm_compass(type = "4star", size = 2, position = c("right", "top")) + 
  tm_scale_bar(position = c("left", "bottom"), text.size = 0.5, breaks = c(0, 0.5, 1)) +
  tm_credits("Created by Laura Hatt", fontface = "bold", 
             position = c("right", "bottom"), size = 0.5) +
  tm_credits("6 March 2023", position = c("right", "bottom"), size = 0.5) 

tmap_save(map_provider_isochrones, paste0(path, "Maps/map_provider_isochrones.png"))
map_provider_isochrones
#add manual legend to label red vs green dots
```

```{r count-families-in-each-isochrone}
num_fams_in_iso <- list()
for (row in seq(1,nrow(prov_isochrones_df), by=1)) {
  fams_in_iso <- lengths(st_intersects(fams_as_pts, prov_isochrones_df[row,])) > 0 
  fams_in_iso <- fams_as_pts[fams_in_iso,]
  num_fams <- nrow(fams_in_iso)
  num_fams_in_iso <- append(num_fams_in_iso, num_fams)
}

num_fams_in_iso <- unlist(num_fams_in_iso)
prov_isochrones_df['num_fams'] <- num_fams_in_iso
```

Note that I assume 1.94 young children per family, per:
https://www.statista.com/statistics/718084/average-number-of-own-children-per-family/#:~:text=The%20typical%20American%20picture%20of,18%20per%20family%20in%201960.&text=If%20there's%20one%20thing%20the,is%20known%20for%2C%20it's%20diversity.

```{r estimate-slot-to-population-ratio-for-each-provider}
prov_isochrones_df['num_child'] <- round(st_set_geometry(prov_isochrones_df, NULL)['num_fams']*1.94,0)
prov_isochrones_df['SPR'] <- st_set_geometry(prov_isochrones_df, NULL)['DyCpcty'] / 
  st_set_geometry(prov_isochrones_df, NULL)['num_child']
```


```{r}
daycare_SPR <- merge(daycare_lakeview, 
                     st_set_geometry(prov_isochrones_df, NULL),
                     by.x = 'PrvdrID', by.y = 'PrvdrID', all.x = TRUE) %>%
  select(PrvdrID, DyCpcty.x, SPR)
```

```{r}
write_sf(obj=daycare_SPR, dsn=paste0(path, "Shapefiles/daycare_SPR.shp"))
```
