==> 5. family-isochrones.Rmd <==
---
title: "family-isochrones"
author: "Laura Hatt"
date: '2023-03-05'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load-packages, warning = FALSE}
library(sf)
library(tmap)
library(tidyverse)

dev_lib_path <-"/Users/laurahatt/Documents/R/dev-versions/"
library(osrm, lib.loc = dev_lib_path)

path = "/Users/laurahatt/Documents/GitHub/child-care-deserts/"
```

```{r}
lakeview <- st_read(paste0(path, "Inputs/comm_areas.geojson")) %>% 
  st_transform("EPSG:3435") %>%
  filter(community == "LAKE VIEW")

#recall that this df includes all families within a 1.5 mile buffer of lakeview
fams_as_pts_lvbuff <- read_sf(paste0(path, "Shapefiles/families_as_points_lakeview.shp"))

#recall that this df includes all daycares within a 1 mile buffer of lakeview
daycare_SPR <- read_sf(paste0(path, "Shapefiles/daycare_SPR.shp"))
```


```{r select-families-in-lakeview-proper, warning = FALSE}
fams_as_pts_lvprop <- st_intersection(fams_as_pts_lvbuff, lakeview) %>% select(FID)
```


note that I'm not considering age range, day vs night, or language


This takes about 20 seconds per family and there are 3000 families. 
If I calculated supply for every family, it would take 15 hours to run. 
Instead, I calculated supply for every tenth family, which took 1.5 hours.
This code is commented out for now so that I don't accidentally run it again.

```{r calculate-child-care-supply-facing-each-family, warning = FALSE}
# fam_isochrones_by10 <- list()
# for (row in seq(1, nrow(fams_as_pts_lvprop), by=10)) {
#   fam_iso <- osrmIsochrone(loc = fams_as_pts_lvprop[row,], breaks = 15, osrm.profile="foot")
#   provs_in_fam_iso <- st_intersection(daycare_SPR, fam_iso)
#   fam_iso['AdjSup'] <- sum(st_set_geometry(provs_in_fam_iso, NULL)['SPR'])
#   fam_iso['FID'] <- st_set_geometry(fams_as_pts_lvprop[row,], NULL)['FID']
#   fam_isochrones_by10[[row]] <- fam_iso
#   print(row)
# }
# 
# fam_isochrones_by10_df = do.call(rbind, fam_isochrones_by10)
# fam_isos <- st_set_geometry(fam_isochrones_by10_df, NULL)
# fam_supply <- merge(fams_as_pts_lvprop, fam_isos, by.x = 'FID', by.y = 'FID', all.x = TRUE)
# fam_supply_by10 <- na.omit(fam_supply)
# 
# write_sf(obj=fam_supply_by10, dsn=paste0(path, "Shapefiles/fam_supply_by10.shp"))

fam_supply_by10 <- read_sf(paste0(path, "Shapefiles/fam_supply_by10.shp"))
```

```{r plot-supply-in-point-form}
tmap_mode("plot")

#set bounding box
bbox_new <- st_bbox(st_buffer(lakeview, 5280)) 
xrange <- bbox_new$xmax - bbox_new$xmin 
yrange <- bbox_new$ymax - bbox_new$ymin 
bbox_new[1] <- bbox_new[1] + (0.1 * xrange) # xmin - left
bbox_new[3] <- bbox_new[3] + (0.1 * xrange) #xmax - right
bbox_new[2] <- bbox_new[2] + (0.1 * yrange) #ymin - bottom
bbox_new <- bbox_new %>% st_as_sfc() 

#rename DyCpcty_x column, because tmap can't manually set the legend title
#when I'm showing variation by dot size instead of dot color
daycare_SPR <- rename(daycare_SPR, Slots = DyCpcty_x)

#round supply to two decimals, to streamline plot legend
fam_supply_by10$AdjSupRd <- round(fam_supply_by10$AdjSup, 2)

map_slots_per_child_point <- tm_shape(st_buffer(lakeview, 5280), bbox=bbox_new) +
  tm_borders(col="white") +
  tm_shape(lakeview) + tm_polygons(alpha=0.5) +
  tm_shape(daycare_SPR) + tm_dots(size="Slots")  +
  tm_shape(fam_supply_by10) + tm_dots(col = "AdjSupRd",
                                      palette = "RdBu",
                                      title = "Slots Per Tot",
                                      legend.hist = FALSE) +
  tm_layout(main.title = "Child Care Availability in Lakeview: \nSlots Per Tot (Point Format)",
            main.title.size = 1.25,
            main.title.position = "center",
            legend.position = c("right", "center"))+ 
  tm_compass(type = "4star", size = 2, position = c("right", "top")) + 
  tm_scale_bar(position = c("left", "bottom"), text.size = 0.5, breaks = c(0, 0.5, 1)) +
  tm_credits("Created by Laura Hatt", fontface = "bold", 
             position = c("right", "bottom"), size = 0.5) +
  tm_credits("6 March 2023", position = c("right", "bottom"), size = 0.5) 
tmap_save(map_slots_per_child_point, paste0(path, "Maps/map_slots_per_child_point.png"))

map_slots_per_child_point
```


```{r}
map_zero_slots_point <- tm_shape(st_buffer(lakeview, 5280), bbox=bbox_new) +
  tm_borders(col="white") +
  tm_shape(lakeview) + tm_polygons(alpha=0.5) +
  tm_shape(daycare_SPR) + tm_dots(size="Slots")  +
  tm_shape(fam_supply_by10[fam_supply_by10$AdjSup == 0,]) + 
  tm_dots(col = "darkred", title = " ", legend.hist = FALSE) +
  tm_layout(main.title = "Child Care Availability in Lakeview: \nRegions with Zero Slots (Point Format)",
            main.title.size = 1.25,
            main.title.position = "center",
            legend.position = c("right", "center"))+ 
  tm_compass(type = "4star", size = 2, position = c("right", "top")) + 
  tm_scale_bar(position = c("left", "bottom"), text.size = 0.5, breaks = c(0, 0.5, 1)) +
  tm_credits("Created by Laura Hatt", fontface = "bold", 
             position = c("right", "bottom"), size = 0.5) +
  tm_credits("6 March 2023", position = c("right", "bottom"), size = 0.5) 

tmap_save(map_zero_slots_point, paste0(path, "Maps/map_zero_slots_point.png"))
map_zero_slots_point
```
