==> 6. interpolating_supply.Rmd <==
---
title: "6. interpolating supply"
author: "Laura Hatt"
date: '2023-03-05'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

CAP doesn't interpolate, they just show the dots
the paper also does not interpolate - they calculate summary statistics
however, I want to interpolate

```{r, warning = FALSE}
library(sf)
library(tmap)
library(tidyverse)

path = "/Users/laurahatt/Documents/GitHub/child-care-deserts/"
```

```{r load-datasets}
lakeview <- st_read(paste0(path, "Inputs/comm_areas.geojson")) %>% 
  st_transform("EPSG:3435") %>%
  filter(community == "LAKE VIEW")

#Recall that this dataset includes every tenth family in Lakeview proper
fam_supply_by10 <- read_sf(paste0(path, "Shapefiles/fam_supply_by10.shp"))
```

```{r interpolate-supply-measurements-using-voronoi-polygons, warning = FALSE}
fam_voro_df <- fam_supply_by10

#creating voronoi polygons 
voro_polygons <- st_voronoi(do.call(c, st_geometry(fam_voro_df))) %>%
  st_collection_extract() %>%
  st_set_crs("EPSG:3435")

#re-ordering voronoi polygons to match order of family points
voro_polygons <- voro_polygons[unlist(st_intersects(fam_supply_by10, voro_polygons))]

#adding voronoi polygon geometry and removing point geometry column
fam_voro_df$voro_polygons <- voro_polygons
fam_voro_df <- st_set_geometry(fam_voro_df, NULL) %>% 
  select(FID, AdjSup, voro_polygons) %>%
  st_as_sf()

#clip voronoi polygons to neighborhood boundaries
fam_voro_df <- st_intersection(fam_voro_df, lakeview) %>%
  select(FID, AdjSup)
```


```{r combine-voronoi-polygons-with-equal-supply}
fam_voro_df$AdjSupRd <- round(fam_voro_df$AdjSup, 2)

fam_voro_comb <- fam_voro_df %>% 
  group_by(AdjSupRd) %>%
  summarize()
```


```{r}
tmap_mode("plot")

#expand bounding box to make room for map elements
#https://www.jla-data.net/eng/adjusting-bounding-box-of-a-tmap-map/
# bbox_new <- st_bbox(lakeview) 
# xrange <- bbox_new$xmax - bbox_new$xmin 
# yrange <- bbox_new$ymax - bbox_new$ymin 
# bbox_new[3] <- bbox_new[3] + (0.4 * xrange) #xmax - right
# bbox_new[2] <- bbox_new[2] - (0.2 * yrange) #ymin - bottom
# bbox_new <- bbox_new %>% st_as_sfc() 

bbox_new <- st_bbox(st_buffer(lakeview, 5280)) 
xrange <- bbox_new$xmax - bbox_new$xmin 
yrange <- bbox_new$ymax - bbox_new$ymin 
bbox_new[1] <- bbox_new[1] + (0.1 * xrange) # xmin - left
bbox_new[3] <- bbox_new[3] + (0.1 * xrange) #xmax - right
bbox_new[2] <- bbox_new[2] + (0.1 * yrange) #ymin - bottom
bbox_new <- bbox_new %>% st_as_sfc() 

map_slots_per_child_voronoi <- tm_shape(lakeview, bbox=bbox_new) + tm_polygons() +
  tm_shape(fam_voro_comb) + tm_polygons(col="AdjSupRd", 
                                           palette="RdBu",
                                           title="Slots Per Tot") +
  tm_layout(main.title = "Child Care Availability in Lakeview: \nSlots Per Tot (Voronoi Tessellation)",
            main.title.size = 1.25,
            main.title.position = "center",
            legend.position = c("right", "center")) +
  tm_compass(type = "4star", size = 2, position = c("right", "top")) + 
  tm_scale_bar(position = c("left", "bottom"), text.size = 0.5, breaks = c(0, 0.5, 1)) +
  tm_credits("Created by Laura Hatt", fontface = "bold", position = c("right", "bottom"), size = 0.5) +
  tm_credits("6 March 2023", position = c("right", "bottom"), size = 0.5)
tmap_save(map_slots_per_child_voronoi, paste0(path, "Maps/map_slots_per_child_voronoi.png"))

map_slots_per_child_voronoi
```

In order to define a "child care desert", one must define some kind of threshold. There is no consensus about this threshold, especially in terms of walking time.

```{r}
#The mean is 0.0271 slots per child.
mean(st_set_geometry(fam_voro_df['AdjSup'], NULL)$AdjSup)

#46.85% of families do not have any licensed childcare providers 
#within a 15-minute walk of their house
ecdf(st_set_geometry(fam_voro_df, NULL)$AdjSup)(0)

#the percentiles are as follows
quantile(st_set_geometry(fam_voro_df, NULL)$AdjSup, probs = seq(0, 1, by= 0.1)) 
```


```{r}
#transforming supply into binary measure: zero or nonzero
fam_voro_df$nonzero <- factor(ifelse(fam_voro_df$AdjSup==0,"Zero","More than zero"),
                             levels = c("Zero", "More than zero"))

#combining Voronoi polygons with the same value
fam_voro_nonzero <- fam_voro_df %>% 
  group_by(nonzero) %>%
  summarize()

map_zero_slots_voronoi <- tm_shape(lakeview, bbox=bbox_new) + tm_polygons() +
  tm_shape(fam_voro_nonzero) + tm_polygons(col="nonzero", 
                                           palette="RdBu",
                                           title="Slots Per Tot") +
  tm_layout(main.title = "Child Care Availability in Lakeview: \nRegions with Zero Slots (Voronoi Tessellation)",
            main.title.size = 1.25,
            main.title.position = "center", 
            legend.position = c("right", "center")) + 
  tm_compass(type = "4star", size = 2, position = c("right", "top")) + 
  tm_scale_bar(position = c("left", "bottom"), text.size = 0.5, breaks = c(0, 0.5, 1)) +
  tm_credits("Created by Laura Hatt", fontface = "bold", position = c("right", "bottom"), size = 0.5) +
  tm_credits("6 March 2023", position = c("right", "bottom"), size = 0.5)
tmap_save(map_zero_slots_voronoi, paste0(path, "Maps/map_zero_slots_voronoi.png"))

map_zero_slots_voronoi
```
